+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Added `longestLength` method in `assignment_01.cpp` and `assignment_01_generic.cpp`.
+- Added `Assignment_02_Binary_Tree` and `linkedqueue.cpp` and `linkedstack.cpp`